Push just enough into the package to test it
