+++ src/pincushion/flask/views.py
@@ -1,0 +1,17 @@
+# -*- encoding: utf-8
+
+from flask import render_template
+
+from . import app
+
+
+@app.errorhandler(404)
+def page_not_found(error):
+    message = (
+        'The requested URL was not found on the server. If you entered the '
+        'URL manually please check your spelling and try again.'
+    )
+    return render_template(
+        'error.html',
+        title='404 Not Found',
+        message=message), 404
