+++ tests/test_flask_filters.py
@@ -47,3 +47,11 @@ def test_custom_tag_sort_ignores_non_wc_tags(tags):
     assume(not any(t.startswith('wc:') for t in tags))
     result = filters.custom_tag_sort(tags)
     assert result == sorted(tags)
+
+
+@pytest.mark.parametrize('query, expected_display_query', [
+    ('Hello world', 'Hello world'),
+    ('They said "Hello"', 'They said &quot;Hello&quot;'),
+])
+def test_display_query(query, expected_display_query):
+    assert filters.display_query(query) == expected_display_query
